Remove unused id_interface functions
id_interface's clone() and deep_copy() are hardly used, it would be nice to get rid of them.
There seems to be one issue, in perform_lambda_calc(); this is the only ref() call / id_ref
usage, which may have become unusable.

On my machine two integration tests are failing for some reason (also on the previously
pushed commit, that passes on GitHub).