Type-erase caching level and introspective
The caching level and introspection flag are no longer part of the function_request type,
but are now part of the function_request_impl type.
The consequence is that a request tree can contain requests with different caching levels
and/or introspection. The only restriction left is that all requests should have the same
type of function: plain or coroutine.